Number of pts in pt clouds counted with int64 type
This is to ensure large points clouds can still be counted properly

+++ io/las_pointcloud_io.cpp
@@ -723,7 +723,7 @@ std::vector<std::string> LasPointCloudHeader::attributeList() const {
     return publicHeaderBlock.publicHeaderAttributeList();
 }
 
-int LasPointCloudHeader::expectedNumberOfPoints() const {
+int64_t LasPointCloudHeader::expectedNumberOfPoints() const {
     return publicHeaderBlock.numberOfPointRecords;
 }
 LasExtraAttributesInfos LasPointCloudHeader::getPointwiseExtraAttributesInfos(bool ignoreUndocumentedExtraBytes) const {
